Move menu shortcut detection and application out of AppActions and into OsDetection
